+++ src/main/java/subway/controller/LineMenu.java
@@ -1,19 +1,23 @@
-package subway.view.menu;
+package subway.controller;
 
 import subway.view.InputView;
 import subway.view.OutputView;
 
-public class SectionMenu {
+public class LineMenu {
     public static void print() {
-        OutputView.printSectionMenu();
-        String selection = InputView.receiveMenu("Section");
+        OutputView.printLineMenu();
+        String selection = InputView.receiveMenu("Line");
 
         if (selection.equals("1")) {
-            // 구간 등록
+            // 노선 등록
             return;
         }
         if (selection.equals("2")) {
-            // 구간 삭제
+            // 노선 삭제
+            return;
+        }
+        if (selection.equals("3")) {
+            // 노선 조회
             return;
         }
         if (selection.equals("B")) {
