+++ src/main/java/subway/controller/exception/NameFormatException.java
@@ -1,4 +1,4 @@
-package subway.view.validator;
+package subway.controller.exception;
 
 public class NameFormatException extends RuntimeException {
     public NameFormatException(String message) {
